requirement.txt
openai
colorama
python-dotenv